Added some confirmation dialog and handling some error with Toastr

+++ src/app/user/user-list/user-list.component.ts
@@ -1,4 +1,5 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {ConfirmationService, ConfirmEventType, MessageService} from "primeng/api";
 import {DialogService} from "primeng/dynamicdialog";
 import {User} from "../../../model/User";
 import {UsersService} from "../../services/users.service";
@@ -20,7 +21,9 @@ export class UserListComponent implements OnInit {
 
   constructor(
     private _usersService: UsersService,
-    public dialogService: DialogService
+    public dialogService: DialogService,
+    private _messageService: MessageService,
+    private _confirmationService: ConfirmationService
   ) {
   }
 
@@ -34,27 +37,53 @@ export class UserListComponent implements OnInit {
     this._usersService.getAllUsers().subscribe(
       users => {
         this.users = users;
-        console.log(this.users);
         this.isLoading = false;
       }, error => {
-        alert('error al obtener los datos');
+        if (error.status == 401) {
+          this._messageService.add({
+            severity: 'error',
+            summary: 'Error',
+            detail: 'No estas autoizado para ver estos datos'
+          });
+          this.isLoading = false;
+        } else {
+          this._messageService.add({
+            severity: 'error',
+            summary: 'Error',
+            detail: `Ha ocurrido un error al obtener datos`
+          });
+          this.isLoading = false;
+        }
       }
     );
   }
 
   onDeleteUser(userId: string) {
-    console.log('userId: ', userId);
-    this._usersService.deleteUser(userId).subscribe({
-      next: (response) => {
-        this._usersService.usersChanged.next();
+    this._confirmationService.confirm({
+      message: 'Quieres eliminar este usuario?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this._usersService.deleteUser(userId).subscribe({
+          next: (response) => {
+            this._usersService.usersChanged.next();
+          },
+          error: (error) => {
+          }
+        });
+        this._messageService.add({severity: 'info', summary: 'Confirmed', detail: 'Usuario eliminado'});
       },
-      error: (error) => {
-      }
+      reject: () => {
+        this._messageService.add({severity: 'warn', summary: 'Cancelada', detail: 'Operación cancelada'});
+      },
+      acceptLabel: 'Si',
+      rejectLabel: 'No'
     });
+
+
   }
 
   openUpdateModal(userId: string) {
-
     this.dialogService.open(UpdateUserModalComponent, {
       data: {
         userId
